refactor: replace ChargeOptimizationSettingsActivity with BatteryDetailsActivity for battery info display
